refactor: 안쓰는 모듈 삭제, svgs를 index에서 가져오도록 수정
